+++ src/main/kotlin/com/flyn/eeg_receiver/event/EEGPowerEvent.kt
@@ -1,6 +1,7 @@
 package com.flyn.eeg_receiver.event
 
-class EEGPowerEvent(val delta: Int,
+class EEGPowerEvent(val time: Long,
+                    val delta: Int,
                     val theta: Int,
                     val lowAlpha: Int,
                     val highAlpha: Int,