+++ src/main/kotlin/com/flyn/eeg_receiver/event/RawWaveEvent.kt
@@ -1,4 +1,4 @@
 package com.flyn.eeg_receiver.event
 
-class RawWaveEvent(val value: Int): DeviceEvent {
+class RawWaveEvent(val time: Long, val value: Int): DeviceEvent {
 }