+++ src/main/kotlin/com/flyn/eeg_receiver/event/StrengthEvent.kt
@@ -1,4 +1,4 @@
 package com.flyn.eeg_receiver.event
 
-class StrengthEvent(val value: Int): DeviceEvent {
+class StrengthEvent(val time: Long, val value: Int): DeviceEvent {
 }