+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.3.0] - 2021-10-08
 ### Added
 - EEG raw data viewer
+- event record what time receive it
 ### Changed
 - move all file to com.flyn.eeg_receiver folder
 - MindSetReceiver class renamed EEGReceiver class